+++ oap-server/server-library/library-elasticsearch-client/src/main/java/org/apache/skywalking/library/elasticsearch/requests/factory/v6/V6DocumentFactory.java
@@ -198,4 +198,23 @@ final class V6DocumentFactory implements DocumentFactory {
                       .content(MediaType.JSON, content)
                       .build();
     }
+
+    @Override
+    public HttpRequest deleteById(final String index, final String type, final String id, Map<String, ?> params) {
+        checkArgument(!isNullOrEmpty(index), "index cannot be null or empty");
+        checkArgument(!isNullOrEmpty(type), "type cannot be null or empty");
+        checkArgument(!isNullOrEmpty(id), "id cannot be null or empty");
+
+        final HttpRequestBuilder builder = HttpRequest.builder();
+        if (params != null) {
+            params.forEach(builder::queryParam);
+        }
+
+        return builder
+            .delete("/{index}/{type}/{id}")
+            .pathParam("index", index)
+            .pathParam("type", type)
+            .pathParam("id", id)
+            .build();
+    }
 }