+++ docs/en/changes/changes.md
@@ -5,6 +5,7 @@
 
 #### OAP Server
 
+* ElasticSearchClient: Add `deleteById` API.
 
 
 #### UI